Enforce the requirement for a "default" reward.
Default rewards have always been required, but are frequently
misconfigured. This is an attempt to resolve the problem in a
user-friendly and backwards-compatible way.

A reward with a if block containing "default: true" or is empty is
considered a "default" reward. Only one reward can be defined this way.
If no rewards (or more than one) reward is defined this way, SuperbVote
will display an error.

Since users often don't realize that an empty if section is equivalent
to a default reward, this commit also adds support for a "default: true"
matcher that makes it very obvious upon visual examination that this is
the default reward given to players.